+++ src/shellyDevice.test.ts
@@ -7,7 +7,7 @@ const HOMEDIR = path.join('jest', NAME);
 import path from 'node:path';
 
 import { jest } from '@jest/globals';
-import { flushAsync, loggerLogSpy, setupTest } from 'matterbridge/jestutils';
+import { flushAsync, loggerLogSpy, setDebug, setupTest } from 'matterbridge/jestutils';
 import { AnsiLogger, BLUE, db, dn, er, hk, LogLevel, MAGENTA, nf, nt, TimestampFormat, YELLOW, zb } from 'matterbridge/logger';
 import { wait } from 'matterbridge/utils';
 
@@ -228,6 +228,30 @@ describe('Shelly devices test', () => {
     device.thermostatSystemModeTimeout = setTimeout(() => {}, 1000); // Prevent timeout in tests
     device.thermostatSetpointTimeout = setTimeout(() => {}, 1000); // Prevent timeout in tests
     (device as any).startWsClientTimeout = setTimeout(() => {}, 1000); // Prevent timeout in tests
+    device.destroy();
+  });
+
+  test('create a dummy btHome gateway device and fetch', async () => {
+    const device = await ShellyDevice.create(shelly, log, path.join('src', 'mock', 'shellydummy-AABBCCDDEEFF.json'));
+    expect(device).toBeDefined();
+    if (!device) return;
+    expect(device.name).toBe('2PM Gen3 Cover');
+    expect(device.id).toBe('shellydummy-AABBCCDDEEFF');
+    expect(device.host).toBe(path.join('src', 'mock', 'shellydummy-AABBCCDDEEFF.json'));
+    expect(device.username).toBe('admin');
+    expect(device.password).toBe('tango');
+    expect(device.components).toHaveLength(13);
+    // @ts-expect-error componentsPayload is private
+    expect(device.componentsPayload.components).toHaveLength(24);
+    expect(device.bthomeTrvs.size).toBe(0);
+    expect(device.bthomeDevices.size).toBe(5);
+    expect(device.bthomeSensors.size).toBe(19);
+
+    const payload = await device.fetchUpdate();
+    expect(payload).not.toBeNull();
+    expect(payload).toBeInstanceOf(Object);
+    expect(payload).toHaveProperty('cover:0');
+
     device.destroy();
   });
 
@@ -548,7 +572,22 @@ describe('Shelly devices test', () => {
 
     // Mock fetch to fail for settings
     fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
-      if (service === 'shelly') return Promise.resolve({});
+      if (service === 'shelly')
+        return Promise.resolve({
+          name: '2PM Gen3 Cover',
+          id: 'shelly2pmg3-34cdb0770c4c',
+          mac: '34CDB0770C4C',
+          slot: 1,
+          model: 'S3SW-002P16EU',
+          gen: 3,
+          fw_id: '20250520-083748/1.6.2-gc8a76e2',
+          ver: '1.6.2',
+          app: 'S2PMG3',
+          auth_en: false,
+          auth_domain: null,
+          profile: 'cover',
+          matter: false,
+        });
       if (service === 'Shelly.GetConfig') return Promise.resolve(null);
       return Promise.resolve({});
     });
@@ -569,7 +608,22 @@ describe('Shelly devices test', () => {
 
     // Mock fetch to fail for status
     fetchSpy.mockImplementation((shelly: Shelly, log: AnsiLogger, host: string, service: string, params?: Record<string, string | number | boolean | object>) => {
-      if (service === 'shelly') return Promise.resolve({});
+      if (service === 'shelly')
+        return Promise.resolve({
+          name: '2PM Gen3 Cover',
+          id: 'shelly2pmg3-34cdb0770c4c',
+          mac: '34CDB0770C4C',
+          slot: 1,
+          model: 'S3SW-002P16EU',
+          gen: 3,
+          fw_id: '20250520-083748/1.6.2-gc8a76e2',
+          ver: '1.6.2',
+          app: 'S2PMG3',
+          auth_en: false,
+          auth_domain: null,
+          profile: 'cover',
+          matter: false,
+        });
       if (service === 'Shelly.GetConfig') return Promise.resolve({});
       if (service === 'Shelly.GetStatus') return Promise.resolve(null);
       return Promise.resolve({});